Creating a title fragment with navigation
Adding a home activity title fragment that has a button to start the
game. Added this new fragment to activity_main. Updated activity main to
include nav host fragment for title fragment which has navGraph and
defaultNavHost (in case you have multiple nav hosts). Added navhost
fragment, title fragment and game fragment with an action from title to
game fragment and updated the title fragment to add the on click
listener that triggers the navgiation to the home fragment