Adding share functionality support
This uses a menu item with the share icon and shows it on the action bar
if there's space. the menu item is winner_menu.xml. Updating gamewon
fragment to include getshareintent(), shareSuccess() which starts the
new activity with intent of sending/sharing with the appropriate text.
FInally, we inflate the menu and add on item selected option.