Creating material Side nav drawer / hamburger menu
Adding the gradle dependency to support nav drawer. Updating the main
activity xml file to add the drawer layout container, and navigation
view that supports the animations and interactions related to nav the
drawer. Creating two menu items that navigate to the rules and about
fragments. Updating MainActivity.java to support the nav with
NavigationUI.